+++ EX06/src/add.c
@@ -1,5 +1,3 @@
 #include "../inc/calc.h"
 
-int add(int a, int b) {
-    return a + b;
-}
+float add(float a, float b) { return a + b; }
