+++ EX06/src/div.c
@@ -1,5 +1,3 @@
 #include "../inc/calc.h"
 
-int div(int a, int b) {
-    return a / b;
-}
+float div(float a, float b) { return a / b; }
