+++ EX06/src/mul.c
@@ -1,5 +1,3 @@
 #include "../inc/calc.h"
 
-int mul(int a, int b) {
-    return a * b;
-}
+float mul(float a, float b) { return a * b; }
