+++ EX06/src/sub.c
@@ -1,5 +1,3 @@
 #include "../inc/calc.h"
 
-int sub(int a, int b) {
-    return a -b;
-}
+float sub(float a, float b) { return a - b; }
